Fixes Date operation
